Move logic to HostInfo and fix region ignored warning and tests.
