+++ src/SonarScanner.MSBuild.PreProcessor/Resources.resx
@@ -616,4 +616,7 @@ Check the certificates in the truststore file '{1}' specified via the {2} parame
   <data name="ERROR_UnsupportedRegion" xml:space="preserve">
     <value>Unsupported region '{0}'. List of supported regions: 'us'. Please check the 'sonar.region' property.</value>
   </data>
+  <data name="WARN_RegionIsOverriden" xml:space="preserve">
+    <value>The {0} parameter is set to "{1}". The setting will be overriden by one or more of the properties {2}, {3}, or {4}.</value>
+  </data>
 </root>
